Move shell to internals
